refactor: theme 파일 분리 및 style 폴더 생성
