+++ src/styles/theme.js
@@ -1,0 +1,9 @@
+const theme = {
+  placeholderColor: "#a5a5a5",
+  woowagreen: "#6FA189",
+  woowalightgreen: "#95DAC5",
+  woowawhite: "#FFFFFF",
+  woowablack: "#000000",
+};
+
+export default theme;
